Fix settings issues
