+++ Game/Assets/Editor/CodeGenerationSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f3a150c4d6af29844ac3d3d7fa4a74c3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
